+++ webgraph/src/utils/par_sort_iters.rs
@@ -227,7 +227,7 @@ impl ParSortIters {
     /// Sorts the output of the provided sequence of iterators of (labelled)
     /// pairs, returning a [`SplitIters`] structure.
     ///
-    /// This method accept as type parameter a
+    /// This method accepts as type parameter a
     /// [`BitSerializer`](crate::traits::BitSerializer) and a
     /// [`BitDeserializer`](crate::traits::BitDeserializer) that are
     /// used to serialize and deserialize the labels.